added option to hide/show internal representations
Internal representation is a textual representation of the object returned by language runtime. In case of Python it is a result of calling repr(obj) or unicode(obj). Or in case of Ruby it would be obj.inspect or obj.to_s. This is what you usually see in a classic text log output. Previously all internal representations were visible in FireLogger as "_" properties. Now they are hidden by default and you can show them with this option.